fix: Address change requests from Cursor review
